fix: move background layer utils to seperate files
